app/test-crypto-perf: add throughput OOP decryption
During throughput running, re-filling the test data will
impact the performance test result. So for now, to run
decrypt throughput testing is not supported since the
test data is not filled.

But if user requires OOP(out-of-place) mode, the test
data from source mbuf will never be modified, and if
the test data can be prepared out of the running loop,
the decryption test should be fine.

This commit adds the support of out-of-place decryption
testing for throughput.

Signed-off-by: Suanming Mou <suanmingm@nvidia.com>
Acked-by: Brian Dooley <brian.dooley@intel.com>