net/ntnic: add generic NIM and I2C modules
As the ntnic can support different port speeds, it also needs to support
different NIMs (Network Interface Module). This commit add the generic
NIM support for ntnic, such that the specific modules, such as QSFP28
can be added later.

The communication with NIMs is in the form of I2C, so support for such a
module is added as well.

Additionally a thread is added to control the NIM stat machines.

Signed-off-by: Serhii Iliushyk <sil-plv@napatech.com>
Reviewed-by: Ferruh Yigit <ferruh.yigit@amd.com>